requirements.txt
django==4.0.5
djangorestframework==3.13.0
pyyaml==6.0
requests==2.28.1
django-cors-headers==3.13.0
pylint==2.14.4
pre-commit==2.19.0
black==22.6.0
django-configurations==2.3.2
python-dotenv==0.20.0
django-softdelete==0.9.4
